Changed github action version to node 18.
I was actually running v16 in the action,
but v14 on my development environment.

I decided it was time to upgrade off of v14
because v18 is now the LTS version, and v14 is
in maintinence. After doing that I realized I
should check my portfolio project to make sure
it's built with the same version I develop with,
and it turned out it wasn't.